+++ functions/writeToGoogleSheets.js
@@ -9,8 +9,7 @@ const serviceAccountKey = is_prod
   : './serviceAccountDev.json'
 const serviceAccount = require(serviceAccountKey)
 const moment = require('moment-timezone')
-const { db } = require('./helpers')
-const fetch = require('node-fetch')
+const { db, calculateTotalDistanceFromLocations } = require('./helpers')
 
 const express = require('express')
 const cors = require('cors')
@@ -157,7 +156,7 @@ const writeToGoogleSheets = async () => {
         .map(s => `${s.organization.name} (${s.location.address1})`)
         .join(', ')
 
-      // TODO: implement full total distance
+      //Add Total Distance to rescue
       rescue.total_distance = await calculateTotalDistanceFromLocations(
         rescue_stops.map(
           stop =>
@@ -470,39 +469,6 @@ function differenceInTime(dateStarted, dateEnded) {
   const m = Math.floor((totalTimeSecs % 3600) / 60)
 
   return `${h} hours, ${m} minutes`
-}
-
-const calculateTotalDistanceFromLocations = async (locations = []) => {
-  console.log(
-    'Running calculateTotalDistanceFromLocations with input:',
-    locations
-  )
-  const base_url =
-    'https://maps.googleapis.com/maps/api/distancematrix/json?units=imperial'
-  const API_KEY = ''
-
-  let total_distance = 0
-  let curr_location = locations.shift()
-
-  while (locations.length) {
-    const full_url = `
-      ${base_url}&key=${API_KEY}
-      &origins=${encodeURIComponent(curr_location)}
-      &destinations=${encodeURIComponent(locations[0])}
-    `
-
-    const response = await fetch(full_url).then(res => res.json())
-    const distance = parseFloat(
-      response.rows[0].elements[0].distance.text.split(' ')[0]
-    )
-
-    console.log(distance)
-    total_distance += distance
-    curr_location = locations.shift()
-  }
-
-  console.log(total_distance)
-  return `${total_distance} miles`
 }
 
 const writeToSheets = express()
